build: set node and yarn versions
Previously, the node and yarn version were not locked down and upon
attempting to install the dependencies on a new machine, it changed the
dependency configuration in some way that caused running the project to
fail.

This commit locks down a working node and yarn version, explicitly lists
a required dev dependency, and provides repeatable instructions for
project setup.

SUSE1-44 #resolve