fix: workaround incorrect ButtonDropdown z-index
This commit adds `z-index: 1` to the cluster-header to ensure it stays
above the ButtonDropdown in the `AppLancherCard` page.

This shouldn't be necessary, since the position:sticky in the
app-launcher page's header should make the header appear above its
siblings, but the ButtonDropdown element seems to have the wrong z-index
by default, so we just force it to be below the header. See the
following slack conversation for previous discussion on this:
https://krumware.slack.com/archives/C053W936JAZ/p1691418985551309?thread_ts=1691099848.592889&cid=C053W936JAZ